requirements.txt
discord.py
python-dotenv
google-cloud-logging
